Created a login page.
I made a login page and a good system for storing/displaying notices and error messages, even across pages. Still no styles, but you can register, activate, login, and logout. The index page is only accessible if you are successfully logged in. If so, it has the link to logout.