Added a password recovery page and functionality.
You can now go to the password recovery page and enter in your email. If the email is a user in the database, an insert is made into the password_recovery table and a link is sent to the user that expires in one hour. When they click the link they will be able to change their password, then they will be redirected to the login page and can login normally with their new password.