Added a simple view profile page. Refs #6
The view profile page displays the name, handle, and bio of the user. If the user being viewed is the same as the already logged in user, an edit button is displayed.

Also added CSS for styling span tags as simple buttons. Submit buttons have been changed to look the same as these simple span buttons.